rearrange github issues into their own test file
