+++ test/runtests.jl
@@ -41,5 +41,6 @@ yasfmt(s, i, m; kwargs...) = fmt(s; kwargs..., i = i, m = m, style = YASStyle())
 
 include("default_style.jl")
 include("yas_style.jl")
+include("issues.jl")
 include("options.jl")
 include("config.jl")
